% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-main.R
\name{combine}
\alias{combine}
\title{Combine data from files}
\usage{
combine(path, pattern, Unlist = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, \code{getwd()}.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}
}
\value{
Combined files' input - each file's contents contained in one list.
}
\description{
Combine data from files
}
