% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{complexityCut}
\alias{complexityCut}
\title{Filter cells by complexity cutoff}
\usage{
complexityCut(mat, cutoff = 3000, justNames = FALSE)
}
\arguments{
\item{mat}{a matrix of cells-by-genes containing gene expression data.}

\item{cutoff}{complexity value below which cells are filtered out.}

\item{justNames}{boolean. If TRUE, return only the names of cells that passed the filtering.}
}
\value{
cell IDs whose detected gene number >= 3000 and -- if \code{justNames==FALSE} -- their detected values.
}
\description{
Removes all cells whose detected gene number/complexity value is below the cutoff.
}
