% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preprocess.R
\name{detected}
\alias{detected}
\title{Cell complexity counter}
\usage{
detected(mat, cellcols = TRUE, rank = TRUE)
}
\arguments{
\item{mat}{a matrix of cells-by-genes or genes-by-cells containing gene expression data.}

\item{cellcols}{boolean. The columns (variables) are cells if TRUE or are genes if FALSE.}

\item{rank}{boolean. If TRUE, order cells by # of non-zero genes detected}
}
\value{
complexity value per cell (# of genes detected per cell)
}
\description{
Cell complexity counter
}
