% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{filter_by_jaccard}
\alias{filter_by_jaccard}
\title{Vector Filtering by Jaccard Similarity}
\usage{
filter_by_jaccard(x, out = NULL, jac.cut = 0.75)
}
\arguments{
\item{x}{list of named vectors. Vectors are assumed to be ordered from most to least significant, such that of two vectors with a similarity score higher than the cut-off, the higher-placed vector will be kept and the lower-placed vector will be thrown.}

\item{out}{vectors that passed jaccard test.}

\item{jac.cut}{jaccard value cut off. arg to \code{jaccard_test}.}
}
\value{
list of i) x: names of vectors that passed jaccard test and ii) out: vectors that passed jaccard test.
}
\description{
One of two vectors with too high jaccard similarity is removed. Priority is given to groups that appear higher in the list of vectors, such that the lower-placed vector is removed.
}
