% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{gene_means}
\alias{gene_means}
\title{Matrix Column/Row Means}
\usage{
gene_means(mat, by = 1)
}
\arguments{
\item{mat}{matrix of vars. by. obs.}

\item{by}{integer of 1 or 2. 1 for row means or 2 for column means.}
}
\value{
vector of mean values
}
\description{
Matrix Column/Row Means
}
