% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{genesCut}
\alias{genesCut}
\title{Filter genes by cutoff}
\usage{
genesCut(mat, cutoff = 4)
}
\arguments{
\item{mat}{a matrix of cells-by-genes containing gene expression data.}

\item{cutoff}{gene expression value below which genes are filtered out.}
}
\value{
matrix without the genes/rows whose aggregate expression across cells was too low.
}
\description{
Removes genes whose average expression across cells is below the cutoff.
}
