% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{hclean}
\alias{hclean}
\title{Remove too big too small clusters}
\usage{
hclean(k, n, min = 5, max = 0.5)
}
\arguments{
\item{k}{chr vector of clusters}

\item{n}{number of members in cluster}

\item{min}{cluster size. absolute value}

\item{max}{cluster size. fraction of total n}
}
\value{
remaining clusters. list of chr vectors
}
\description{
Remove too big too small clusters
}
