% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{hcluster}
\alias{hcluster}
\title{Hierarchical clustering}
\usage{
hcluster(mat, method.cor = "pearson", method.hc = "average",
  dissim.dist = 1, ...)
}
\arguments{
\item{mat}{matrix of vars. vs. obs.}

\item{method.cor}{distance metric for calculating correlations.}

\item{method.hc}{type of linkage.}

\item{dissim.dist}{maximum dissimilarity prior to rescaling between 0 and 1.}

\item{...}{other args passed to \code{cor} or \code{hclust} calls.}
}
\value{
hc object
}
\description{
Hierarchical clustering
}
