% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{hcorder}
\alias{hcorder}
\title{Order variables by hierarchical clustering}
\usage{
hcorder(mat)
}
\arguments{
\item{mat}{matrix of vars. vs. obs.}
}
\value{
chr vector of ordered vars.
}
\description{
Order variables by hierarchical clustering
}
