% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{hcorr}
\alias{hcorr}
\title{Calculate (ordered) pairwise correlation matrix}
\usage{
hcorr(mat, method.cor = "pearson", order = TRUE, melt = FALSE, ...)
}
\arguments{
\item{mat}{matrix of vars. vs. obs. to be correlated.}

\item{method.cor}{distance metric for calculating correlations.}

\item{order}{if TRUE, corr object will be returned with vars ordered by hc.}

\item{melt}{if TRUE, input matrix will be converted to tidy/long format.}

\item{...}{other args passed to \code{cor}.}
}
\value{
matrix of vars. vs. vars correlation scores.
}
\description{
Calculate (ordered) pairwise correlation matrix
}
