% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{hcutree}
\alias{hcutree}
\title{Get clusters from hc object by tree heights}
\usage{
hcutree(hc, h = hc$height, clean = TRUE, min = 5, max = 0.5, ...)
}
\arguments{
\item{hc}{hierarchical clustering obj.}

\item{h}{\code{cutree} arg. height at which to cut. hc$height=all heights.}

\item{clean}{logical. if TRUE, call to \code{hclean}.}

\item{min}{hclean arg. cluster size. absolute value.}

\item{max}{hclean arg. cluster size. fraction of total n.}

\item{...}{other args passed to \code{cutree}.}
}
\value{
clusters. list of chr vectors.
}
\description{
Get clusters from hc object by tree heights
}
