% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Similarity Operation}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{character vector}

\item{b}{character vector}
}
\value{
jaccard similarity value
}
\description{
Jaccard Similarity Operation
}
