% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{jaccard_test}
\alias{jaccard_test}
\title{Jaccard Test}
\usage{
jaccard_test(x, y, jac.cut = 0.75)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}

\item{jac.cut}{jaccard value cut off.}
}
\value{
logical. TRUE if vectors' similarity is below or equal to cut off. FALSE otherwise.
}
\description{
Jaccard Test
}
