% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{logTPM}
\alias{logTPM}
\title{Convert expression data from TPM to logTPM form.}
\usage{
logTPM(tpm, dividebyten = TRUE)
}
\arguments{
\item{tpm}{a numeric vector. Either a matrix or of length 1.}

\item{dividebyten}{a boolean. Usually TRUE when single cells expression vals are concerned; usually FALSE when bulk (many cell) expression vals are concerned.}
}
\value{
a numeric vector (either a matrix or of length 1) with values converted from TPM to logTPM form.
}
\description{
TPM aka Transcripts Per Million.
If a matrix is provided, \code{logTPM} will be applied over each cell and returned as a matrix.
Note. \code{logTPM(tpm) == TPM(logtpm)}
}
