% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-main.R
\name{makeFilename}
\alias{makeFilename}
\title{Generate filename to match Cache files}
\usage{
makeFilename(args, pipeName = NULL, fnName = NULL, cachePath = ".",
  sep = ":", collapse = "__", argNames = FALSE)
}
\arguments{
\item{args}{a list of arguments to the main function}

\item{pipeName}{the name of main function}

\item{fnName}{the name of function that is a step in the main function}

\item{cachePath}{a character string providing path to the Cache directory.}

\item{sep}{passed to \code{string}. a character to separate arguments and their names. Defaults to ":".}

\item{collapse}{passed to \code{string}. a character to separate arguments(+names) from one another. Defaults to "__".}

\item{argNames}{passed to \code{string}. Boolean indicating whether argument names should be included in the final output.}
}
\value{
character string that is the generated filename. It includes the main function and step function names, and the arguments.
}
\description{
Generate filename to match Cache files
}
