% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster.R
\name{ncutree}
\alias{ncutree}
\title{Get clusters from hc object by predefined cluster number}
\usage{
ncutree(hc, n)
}
\arguments{
\item{hc}{hierarchical clustering object.}

\item{n}{number of clusters to return.}
}
\value{
clusters. list of chr vectors.
}
\description{
Get clusters from hc object by predefined cluster number
}
