% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{p_2}
\alias{p_2}
\title{Find the most significant occurrence of each observation}
\usage{
p_2(x, y)
}
\arguments{
\item{x}{numeric vector of (named) observations}

\item{y}{numeric vector of (named) observations}
}
\value{
numeric vectors x and y, trimmed, such that each observation that was present in both now appears only in that group where its occurrence was most significant
}
\description{
Find the most significant occurrence of each observation
}
