% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{p_val}
\alias{p_val}
\title{P-values between matrices}
\usage{
p_val(a, b, p.value = NULL, adjust.method = "bonferroni")
}
\arguments{
\item{a}{matrix (vars. by. obs.) for which p-values are calculated.}

\item{b}{matrix (vars. by. obs.) for which the p-values of matrix 'a' are calculated relative to.}

\item{p.value}{NULL or numeric. If NULL, return all p-values. If numeric, return p-values <= p.value.}

\item{adjust.method}{NULL or character string. If NULL, do not adjust p-values. If string, adjust p-values using the method specified.}
}
\value{
numeric vector containing p-values.
}
\description{
P-values between matrices
}
