% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{pairs_apply}
\alias{pairs_apply}
\title{Apply function to every pair between x and a list}
\usage{
pairs_apply(a, vect, FUN = overlap)
}
\arguments{
\item{a}{value or vector}

\item{vect}{list of values or vectors of the same type as 'a'}

\item{FUN}{function to be applied on every pair. eg \code{sum(a, vect[[1]])} if a and vect are numeric}
}
\value{
result of function call for each pair
}
\description{
Apply function to every pair between x and a list
}
