% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pairs_apply_all}
\alias{pairs_apply_all}
\title{Apply function to every pair in a list}
\usage{
pairs_apply_all(vect, FUN = overlap)
}
\arguments{
\item{vect}{a list. All elements must be of the same type.}

\item{FUN}{function to be applied on every pair. eg sum(a, vect[[1]]) if a and vect are numeric}
}
\value{
result of function call for each pair
}
\description{
Apply function to every pair in a list
}
