% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-score.R
\name{program}
\alias{program}
\title{Programs}
\usage{
program(List, cutoff = 50)
}
\arguments{
\item{List}{list of named numeric vectors. Numbers are p-values and names are genes.}

\item{cutoff}{a numeric value indicating the cutoff for program size}
}
\value{
list of character vectors. Each vector is a program
}
\description{
Programs
}
