% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster-significance.R
\name{sig}
\alias{sig}
\title{Significance according to p-values of fold changes.}
\usage{
sig(a, b, p.value, fc.value = 3, fc.sort = F, pval.sort = F,
  adjust.method = NULL, ...)
}
\arguments{
\item{a}{matrix (vars. by. obs.) for which significance (of each obs.) is calculated}

\item{b}{matrix (vars. by. obs.) for which significance (of each obs.) in matrix 'a' is calculated relative to.}

\item{p.value}{arg to \code{p_val}. NULL or numeric. If NULL, return all p-values. If numeric, return p-values <= p.value.}

\item{fc.value}{arg to \code{fold_change}. NULL or numeric. If NULL, return all FC values. If numeric, return FC values >= fc.value.}

\item{fc.sort}{if TRUE, significantly differentially expressed genes are sorted by fold change (highest first). Default is TRUE.}

\item{pval.sort}{if TRUE, significantly differentially expressed genes are sorted by p.value (highest first). pval.sort=TRUE overrides fc.sort=TRUE. Default is FALSE.}

\item{adjust.method}{NULL or character string. If NULL, do not adjust p-values. If string, adjust p-values using the method specified.}

\item{...}{other args passed to \code{fold_change} or \code{p_val}.}
}
\value{
one of: i) all p-values for all fold changes (if fc.value == NULL & p.value == NULL) ii/iii) all/signifcant p-values for significant/all fold changes or iv) signifcant p-values for significant fold changes (if fc.value != NULL & p.value != NULL).
}
\description{
Significance according to p-values of fold changes.
}
