% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{sig_2}
\alias{sig_2}
\title{Trim vectors of observations such that only the most significant occurrence of each observation is kept.}
\usage{
sig_2(List)
}
\arguments{
\item{List}{a list of vectors of observations. E.g. p-values named by the genes they belong to. some genes (observations) will have multiple occurrences in different vectors, and only the vector with the most significant read-out for that gene gets to keep it.}
}
\value{
a list of vectors of observations. Each observation (gene) only appears once.
}
\description{
Trim vectors of observations such that only the most significant occurrence of each observation is kept.
}
