% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-main.R
\name{string}
\alias{string}
\title{Generate one string from a list of strings}
\usage{
string(args, argNames = FALSE, sep = ":", collapse = "_")
}
\arguments{
\item{args}{named list of arguments}

\item{argNames}{boolean indicating whether argument names should be included in the final output.}

\item{sep}{a character to separate arguments and their names. Defaults to ":".}

\item{collapse}{a character to separate arguments(+names) from one another. Defaults to "__".}
}
\value{
string of concatenated arguments
}
\description{
Generate one string from a list of strings
}
