% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-score.R
\name{top_program_genes}
\alias{top_program_genes}
\title{Top Genes across Programs}
\usage{
top_program_genes(Programs, cutoff = 50)
}
\arguments{
\item{Programs}{list of character vectors. Each vector is a program}

\item{cutoff}{a numeric value indicating how many top genes to return.}
}
\value{
table of top genes and their counts across programs
}
\description{
Top Genes across Programs
}
