% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistrics.R
\name{TPM}
\alias{TPM}
\title{Convert expression data from logTPM to TPM form.}
\usage{
TPM(logtpm, dividebyten = TRUE)
}
\arguments{
\item{logtpm}{a numeric vector. Either a matrix or of length 1.}

\item{dividebyten}{a boolean. Usually TRUE when single cells expression vals are concerned; usually FALSE when bulk (many cell) expression vals are concerned.}
}
\value{
a numeric vector (either a matrix or of length 1) with values converted from logTPM to TPM form.
}
\description{
TPM aka Transcripts Per Million.
If a matrix is provided, \code{TPM} will be applied over each cell and returned as a matrix.
Note. \code{TPM(logtpm) == logTPM(tpm)}
}
