% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-main.R
\name{try_apply}
\alias{try_apply}
\title{Title}
\usage{
try_apply(List, cachePath = ".", FUN = "main", args = NULL,
  pipeName = NULL, ...)
}
\arguments{
\item{List}{List of matrices to apply FUN to.}

\item{cachePath}{Output path where files will be saved in / loaded from if they already exist.}

\item{FUN}{function to apply to list of matrices. Defaults to statistrics::main}

\item{args}{a list of arguments that will define the filename.}

\item{pipeName}{a job ID, a name for the project/pipeline.}

\item{...}{other arguments to be passed to FUN (statistrics::main)}
}
\value{
output of FUN, for each matrix in turn
}
\description{
Title
}
