% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster-significance.R
\name{DEgenes}
\alias{DEgenes}
\title{Differentially Expressed Genes Between Two Groups}
\usage{
DEgenes(k, mat, fc.value = 3, p.value = 10^(-4), fc.sort = T,
  pval.sort = F, adjust.method = NULL, ...)
}
\arguments{
\item{k}{character vector of vars, for which a subgroup will have DE genes calculated.}

\item{mat}{matrix of vars. vs. obs. matrix is split into two: i) vars that are in k; ii) vars not in k.}

\item{fc.value}{fold change value below which differential gene expression is deemed insignificant.}

\item{p.value}{p-value above which differential gene expression is deemed insignificant.}

\item{fc.sort}{if TRUE, significantly differentially expressed genes are sorted by fold change (highest first). Default is TRUE.}

\item{pval.sort}{if TRUE, significantly differentially expressed genes are sorted by p.value (highest first). pval.sort=TRUE overrides fc.sort=TRUE. Default is FALSE.}

\item{adjust.method}{}

\item{...}{other args to be passed to \code{p_val} or \code{fold_change} through call to \code{sig}.}
}
\value{
numeric vector of p-values named with gene names.
}
\description{
Differentially Expressed Genes Between Two Groups
}
