% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster-significance.R
\name{hcsig}
\alias{hcsig}
\title{hcsig: hcluster Significance}
\usage{
hcsig(k, mat, fc.value = 3, p.value = 10^(-4), reorder = TRUE,
  fc.sort = T, pval.sort = F)
}
\arguments{
\item{fc.value}{fold change value below which differential gene expression is deemed insignificant.}

\item{p.value}{p-value above which differential gene expression is deemed insignificant.}

\item{reorder}{if TRUE, the list of clusters is reordered by most to least significant.}

\item{fc.sort}{if TRUE, significantly differentially expressed genes are sorted by fold change (highest first). Default is TRUE.}

\item{pval.sort}{if TRUE, significantly differentially expressed genes are sorted by p.value (highest first). \code{pval.sort=TRUE} overrides \code{fc.sort=TRUE}. Default is FALSE.}
}
\description{
hcluster Significance. For clusters derived from hierarchical clustering (with \code{hclust}), data is retrieved.
}
