% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster-similarity.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Similarity}
\usage{
jaccard(x, y, jac.cut = 0.5, test = TRUE)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}

\item{jac.cut}{jaccard value cut off.}

\item{test}{a boolean indicating whether to test jaccard value against cutoff \code{jac.cut}.}
}
\value{
If no cutoff tested, returns jaccard similarity. If cutoff applied, returns TRUE if jaccard similarity is below or equal to cutoff and FALSE otherwise.
}
\description{
Jaccard Similarity
}
