% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cluster-similarity.R
\name{overlap}
\alias{overlap}
\title{Basic Similarity (intersection value)}
\usage{
overlap(a, b)
}
\arguments{
\item{a}{character vector}

\item{b}{character vector}
}
\value{
overlap (intersection) value
}
\description{
Basic Similarity (intersection value)
}
