% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-preprocess.R
\name{preprocess}
\alias{preprocess}
\title{PreProcess Expression Matrix}
\usage{
preprocess(mat, pipeName = "preprocess", cachePath = ".",
  logTransform = TRUE, complexity.cutoff = 3000, genes.cutoff = 4)
}
\arguments{
\item{mat}{matrix of vars. by. obs. (in TPM or logTPM).}

\item{pipeName}{a job ID, a name for the project/pipeline. Defaults to function name.}

\item{cachePath}{passed to \code{cacheCall::cacheCall}. A character string providing path to the Cache directory.}

\item{logTransform}{if TRUE, apply \code{logTPM} to matrix.}

\item{complexity.cutoff}{passed to \code{cutoff} parameter in complexityCut.}

\item{genes.cutoff}{passed to \code{cutoff} parameter in genesCut.}

\item{...}{additional arguments to be passed to \code{logTPM}.}
}
\value{
centered, log-transformed matrix consisting of only high-quality -- user-defined -- cells and genes)
}
\description{
If the data is passed in TPM form, the data is first transformed to logTPM.
Low quality cells are then removed. They are a filtered according to a cutoff for the number of genes detected.
Next, low quality genes are removed. These are filtered according to a cutoff for their avg. expression values across all cells.
Finally, the data is centered such that the average expression of each gene across all cells is 0.
Note: The standard deviation is left as is -- rather than being normalised to equal 1 --
since s.d. values are very skewed due to how sparse the scRNA data is (many 0s).
}
