% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-score.R
\name{score}
\alias{score}
\title{Program Scores}
\usage{
score(mat, programs, center = TRUE, center.rowWise = FALSE)
}
\arguments{
\item{mat}{a matrix of gene expression data (cells by genes)}

\item{programs}{either a single or a list of character vectors (genes). Each vector is a program.}

\item{center}{if TRUE, the resulting score matrix is centered.}
}
\value{
list (if many=F) or matrix (if many=T) of program scores
}
\description{
Program Scores
}
